+++ scripts/validate_hacs_zip.py
@@ -9,6 +9,7 @@ from zipfile import ZipFile
 REQUIRED_ROOT_FILES = {
     "__init__.py",
     "config_flow.py",
+    "icon.png",
     "manifest.json",
     "sensor.py",
     "services.yaml",